[New] Added test cases for tooltip
